+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ## [3.5.1901.0 - unreleased]
 
 ### Added
+- Added CanProvision method to double-check provisioning requirements. [paolopia]
 
 ### Changed